+++ chevah/compat/interfaces.py
@@ -202,6 +202,8 @@ class IOSUsers(Interface):
         Return the name of the first group from `groups` that `username` is
         member of.
 
+        Raise a ValueError if the requested list of `groups` is empty.
+
         Retrun None if `username` is not a member of any of the groups.
         """